traffic/requirements.txt
Flask==0.11.1
numpy==1.11.1
pandas==0.18.0
Pillow==3.4.2
cv2==1.0
tensorflow==0.11.0rc2
